Structure of the S4 class "Organism"

Structure of the S4 class "Organism"

Slots

cellDiameter

Numeric for the initial Diameter of cell spheres in \(\mu\)m

cellMassInit

Numeric for the initial cell mass in pg

cellMassAtDivision

Limit of a cell's mass before it divided into two daughter cells. Unit: pg

cellShape

Character for the cell shape type. Currently, only coccus/sphere shapes are supported.

vmax

Numeric for the maximum speed a cell can move in \(\mu\)m/s

scavengeDist

Numeric indicating the maximum distance (from cell surface) a cell can scavenge nutrients from its surrounding. Unit: \(\mu\)m

chemotaxisCompound

Character vector with the compound IDs that influence the cells chemotaxis behavior.

chemotaxisStrength

Numeric vector that indicates the strength of chemotaxis. Positive value for attraction; Negative for repelling effect. A value of 1 indicates that in case of a maximum gradient (concentration-weighted center in cell's scavenge area is at the edge of the area) the cell moves with its maximum speed (vmax) in the direction of the gradient.

chemotaxisHillKA

Numeric vector for K_A value in Hill equation in chemotactic metabolite sensing. Unit: mM

chemotaxisHillCoef

Numeric vector for the Hill coefficient. Unitless

mod

Object of S4-class modelorg for the organisms metabolic network model.

exoenzymes

Character vector of IDs of the organism's exoenzymes

exoenzymes.prod

Numeric vector of the production rates of exoenzymes. Unit: nmol / gDW / hr (nmol Enzyme per gDW cells per hr)